refactor: move getIncomingTotals to balances store
